third_party/blink/web_tests/wpt_internal/prerender/unload.html

<!DOCTYPE html>
<!--
This file cannot be upstreamed to WPT until:
* The test changes to use a method to trigger cancellation of prerendering that
  is guaranteed by the specification. Currently the test cancels prerendering
  by performing a main frame navigation after the initial prerendering
  navigation.
* Cross-origin subframe navigation is aligned with the spec. Currently the test
  expects the navigation to be deferred until prerendering activation. Also
  document.prerendering state is unexpectedly unstable on resuming navigation.
  (https://crbug.com/1222547)
* `unload` event handling matches what the specification expects. the current
  specification expects that the unload event handler is never fired, and the
  test passes whether or not it is fired.
  prerendering, but it would be practically difficult, and need a spec change.
* `pageshow` and `pagehide` behaviors during prerendering is clearly defined
  in the specification. This test just checks current behaviors.
  (https://crbug.com/1222551)
-->
<title>unload event handlers</title>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
function createGotMessagesPromise(bc) {
  return new Promise(resolve => {
    bc._messages = [];
    bc.addEventListener('message', e => {
      bc._messages.push(e.data);
      if (e.data == 'Done')
        resolve(bc._messages);
    });
  });
}

function startTest(test) {
  const url = `resources/unload.html?state=start&testcase=${test}`;
  window.open(url, '_blank', 'noopener');
}

promise_test(async t => {
  const bc = new BroadcastChannel('prerender-same-origin-subframe-navigation');
  t.add_cleanup(_ => bc.close());

  const gotMessages = createGotMessagesPromise(bc);
  startTest('same-origin-subframe-navigation');
  const result = await gotMessages;

  const expected = [
    'load start',
    'load prerendering',
    'load same-origin-frame1',
    'pagehide same-origin-frame1 in prerendering',
    'unload same-origin-frame1 in prerendering',
    'load same-origin-frame2',
    'Done',
  ];
  assert_array_equals(result, expected, 'gotMessages');
}, 'unload on same origin subframe navigation');

promise_test(async t => {
  const bc = new BroadcastChannel('prerender-cross-origin-subframe-navigation');
  t.add_cleanup(_ => bc.close());

  const gotMessages = createGotMessagesPromise(bc);
  startTest('cross-origin-subframe-navigation');
  const result = await gotMessages;

  const expected = [
    'load start',
    'load prerendering',
    'load same-origin-frame',
    'request activation',
    'visibilitychange same-origin-frame',
    'pagehide same-origin-frame',
    'visibilitychange same-origin-frame',
    'unload same-origin-frame',
    'load cross-origin-frame',
    'Done',
  ];
  assert_array_equals(result, expected, 'gotMessages');
}, 'unload after activation on cross origin subframe navigation');

promise_test(async t => {
  const bc = new BroadcastChannel('prerender-remove-subframe');
  t.add_cleanup(_ => bc.close());

  const gotMessages = createGotMessagesPromise(bc);
  startTest('remove-subframe');
  const result = await gotMessages;

  const expected = [
    'load start',
    'load prerendering',
    'load frame',
    'request removal',
    'pageshow frame in prerendering',
    'pagehide frame in prerendering',
    'unload frame in prerendering',
    'Done',
  ];
  assert_array_equals(result, expected, 'gotMessages');
}, 'unload on removing subframe');

promise_test(async t => {
  const bc = new BroadcastChannel('prerender-main-frame-navigation');
  t.add_cleanup(_ => bc.close());

  const gotMessages = createGotMessagesPromise(bc);
  startTest('main-frame-navigation');
  const result = await gotMessages;

  const expected = [
    // If `optional` is specified as true, that entry may not appear.
    { message: 'load start' },
    { message: 'load prerendering in prerendering' },
    { message: 'request fallback' },
    // TODO(https://crbug.com/1200241): `pagehide` and `unload` may not run
    // sometimes for unknown reason. Clarify the reason and fix it.
    { message: 'pagehide main-frame in prerendering', optional: true },
    { message: 'unload main-frame in prerendering', optional: true },
    { message: 'load prerendering' },
    { message: 'pagehide main-frame' },
    { message: 'visibilitychange main-frame' },
    { message: 'unload main-frame' },
    { message: 'load another-page' },
    { message: 'Done' },
  ];
  let offset = 0;
  assert_less_than_equal(result.length, expected.length);
  for (let i = 0; i < expected.length; ++i) {
    assert_less_than(offset, result.length);
    while (expected[i].optional && expected[i].message != result[offset])
      continue;
    assert_equals(result[offset], expected[i].message, `messages${offset}`);
    offset++;
  }
  assert_equals(offset, result.length);
}, 'unload on main frame navigation to cancel prerendering');
</script>
